fixed a security issue with !remove, implemented a role check for members unfit for the raid, Improved welcome message.

+++ commands/fun/raid.js
@@ -1,20 +1,20 @@
 const { RichEmbed } = require("discord.js");
 const { stripIndents } = require("common-tags");
 const { promptMessage, getMember } = require("../../functions.js");
-const { raid_channel } = require("../../config.json");
+const { raid_channel, unfit_raid } = require("../../config.json");
 
 module.exports = {
     name: "raid",
     category: "fun",
-    description: "adds the raid role",
+    description: "Kicks the member",
     usage: "<id | mention>",
     run: async (client, message, args) => {
         const logChannel = message.guild.channels.find(c => c.name === `${raid_channel}`) || message.channel;
 
         if (message.deletable) message.delete();
 
-       
-
+        let unraid = (message.member.roles.find(p => p.name === `${unfit_raid}`)) 
+        
         const toKick = getMember(message, args.join(" "));
         let role = message.guild.roles.find(r => r.name === "raid")
         if (!role) return message.reply("This role doesn't exist. Maybe add it??");
@@ -33,24 +33,28 @@ module.exports = {
             .setAuthor(`This verification becomes invalid after 30s.`)
             .setDescription(`Do you want to join the raid?`)
 
-        // Send the message
-        await message.channel.send(promptEmbed).then(async msg => {
-            // Await the reactions and the reaction collector
-            const emoji = await promptMessage(msg, message.author, 30, ["✔", "❌"]);
+        if (!unraid) {
+            // Send the message
+            await message.channel.send(promptEmbed).then(async msg => {
+                // Await the reactions and the reaction collector
+                const emoji = await promptMessage(msg, message.author, 30, ["✔", "❌"]);
 
-            // The verification stuffs
-            if (emoji === "✔") {
-                msg.delete();
-                await toKick.addRole(role.id).catch(e => console.log(e.message))
-                message.channel.send(`Welcome to the big boy league :grin:`)
-                
-                logChannel.send(embed);
-            } else if (emoji === "❌") {
-                msg.delete();
+                // The verification stuffs
+                if (emoji === "✔") {
+                    msg.delete();
+                    await toKick.addRole(role.id).catch(e => console.log(e.message))
+                    message.channel.send(`Welcome to the big boy league :grin:`)
 
-                message.reply(`Not ready yet? Maybe next time.`)
-                    .then(m => m.delete(10000));
-            }
-        });
+                    logChannel.send(embed);
+                } else if (emoji === "❌") {
+                    msg.delete();
+
+                    message.reply(`Not ready yet? Maybe next time.`)
+                        .then(m => m.delete(10000));
+                }
+            });
+        } else {
+            return message.reply("You are unfit for the raid. If you think that is wrong, please contact a staff member.")
+        }
     }
 };
